requirements.txt
gym >= 0.21.0, < 0.24.0
numpy >= 1.20.2
flax >= 0.5.2
ml_collections >= 0.1.0
tqdm >= 4.60.0
optax >= 0.1.3
absl-py >= 0.12.0
scipy >= 1.6.0
wandb >= 0.12.14
tensorflow-probability >= 0.17.0
matplotlib == 3.4.3
mujoco >= 2.2.2
moviepy >= 1.0.3
imageio >= 2.21.3
# d4rl >= 1.1 not required for deepsea
plotly >= 5.10.0
pandas >= 2.1.0
distrax >= 0.1.4
gymnax >= 0.0.6
